Rename `app_delegate` to `app_state`
